% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochrane.R
\name{get.review}
\alias{get.review}
\title{Gets the data directly from the Cochrane Library with a html session.}
\usage{
get.review(doi, path, show.terms = TRUE)
}
\arguments{
\item{doi}{doi address (full http URL or short version).}

\item{path}{path to store data file.}

\item{show.terms}{whether to show link to Cochrane's Terms and Conditions.}
}
\description{
Gets the data directly from the Cochrane Library with a html session.
}
