% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochrane.R
\name{parse.ma}
\alias{parse.ma}
\title{Reads only meta-analytic data.}
\usage{
parse.ma(file, path)
}
\arguments{
\item{file}{RevMan (.rm5) file name.}

\item{path}{path where data file is located.}
}
\value{
data frame with data from a meta-analysis per row.
}
\description{
Reads only meta-analytic data.
}
