% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochrane.R
\name{parse.studies}
\alias{parse.studies}
\title{Reads only study data.}
\usage{
parse.studies(file, path)
}
\arguments{
\item{file}{RevMan (.rm5) file name.}

\item{path}{path where data file is located.}
}
\value{
data frame with data from a study per row.
}
\description{
Reads only study data.
}
