% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochrane.R
\name{read.review}
\alias{read.review}
\title{Reads data from a systematic review from a RevMan data file (XML) from the Cochrane Library.}
\usage{
read.review(file, path)
}
\arguments{
\item{file}{filename.}

\item{path}{path to file.}
}
\value{
a list with two objects (data.frames), a table of the study data, and a table of the meta-analytic data.
}
\description{
Reads data from a systematic review from a RevMan data file (XML) from the Cochrane Library.
}
